feat: Introduce CharacterInputFieldSnapshotTests
