+++ wire-ios/Wire-iOS Tests/CharacterInputFieldTests.swift
@@ -199,64 +199,3 @@ final class CharacterInputFieldTests: XCTestCase {
     }
 }
 
-final class CharacterInputFieldScreenshotTests: XCTestCase {
-    var sut: CharacterInputField! = nil
-
-    override func setUp() {
-        super.setUp()
-        let size = CGSize(width: 375, height: 56)
-        sut = CharacterInputField(maxLength: 8, characterSet: CharacterSet.decimalDigits, size: size)
-
-        sut.frame = CGRect(origin: .zero, size: size)
-    }
-
-    override func tearDown() {
-        sut.removeFromSuperview()
-        sut = nil
-        super.tearDown()
-    }
-
-    func testDefaultState() {
-        // then
-        verify(matching: sut)
-    }
-
-    func testFocusedState() {
-        // given
-        UIApplication.shared.firstKeyWindow?.rootViewController?.view.addSubview(sut)
-
-        // when
-        sut.becomeFirstResponder()
-
-        // then
-        verify(matching: sut)
-    }
-
-    func testFocusedDeFocusedState() {
-        // given
-        UIApplication.shared.firstKeyWindow?.rootViewController?.view.addSubview(sut)
-
-        // when
-        sut.becomeFirstResponder()
-        sut.resignFirstResponder()
-
-        // then
-        verify(matching: sut)
-    }
-
-    func testOneCharacterState() {
-        // when
-        sut.insertText("1")
-
-        // then
-        verify(matching: sut)
-    }
-
-    func testAllCharactersEnteredState() {
-        // when
-        sut.insertText("12345678")
-
-        // then
-        verify(matching: sut)
-    }
-}
