+++ wire-ios/Wire-iOS Tests/ReferenceImages/CharacterInputFieldSnapshotTests/testAllCharactersEnteredState.1.png
