+++ wire-ios/Wire-iOS Tests/ReferenceImages/CharacterInputFieldSnapshotTests/testDefaultState.1.png
