+++ wire-ios/Wire-iOS Tests/ReferenceImages/CharacterInputFieldSnapshotTests/testFocusedDeFocusedState.1.png
