+++ wire-ios/Wire-iOS Tests/ReferenceImages/CharacterInputFieldSnapshotTests/testFocusedState.1.png
