+++ wire-ios/Wire-iOS Tests/ReferenceImages/CharacterInputFieldSnapshotTests/testOneCharacterState.1.png
